Restructured project to be better extensible and more readable
